pyproject.toml
[project]
name = "vit-scope"
version = "0.1.0"
description = "A tool for efficiently using interpretability methods on Vision Transformers (ViTs)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "pillow>=10.1.0",
    "numpy>=1.24.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.21.0",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
only-include = ["backend/app"]
